+++ EssentialFeed/EssentialFeediOSTests/FeedSnapshotTests.swift
@@ -94,62 +94,6 @@ class FeedSnapshotTests: XCTestCase {
         ]
     }
     
-    private func assert(snapshot: UIImage, named name: String, file: StaticString = #file, line: UInt = #line) {
-        let snapshotURL = makeSnapshotURL(named: name, file: file)
-        let snapshotData = makeSnapshotData(for: snapshot)
-        
-        guard let storedSnapshotData = try? Data(contentsOf: snapshotURL) else {
-            XCTFail("Failed to load stored snapshot at URL: \(snapshotURL). Use the `record` method to store a snapshot before asserting.", file: file, line: line)
-            return
-        }
-        
-        if snapshotData != storedSnapshotData {
-            let temporarySnapshotURL = URL(fileURLWithPath: NSTemporaryDirectory(), isDirectory: true)
-                .appendingPathComponent(snapshotURL.lastPathComponent)
-            
-            try? snapshotData?.write(to: temporarySnapshotURL)
-            
-            XCTFail("New snapshot does not match stored snapshot. New snapshot URL: \(temporarySnapshotURL), Stored snapshot URL: \(snapshot)", file: file, line: line)
-        }
-    }
-    
-    private func record(snapshot: UIImage, named name: String, file: StaticString = #file, line: UInt = #line) {
-        let snapshotURL = makeSnapshotURL(named: name, file: file)
-        let snapshotData = makeSnapshotData(for: snapshot)
-        
-        // Next, we have to make sure that this folder structure exists in the File system.
-        do {
-            try FileManager.default.createDirectory(
-                at: snapshotURL.deletingLastPathComponent(),
-                withIntermediateDirectories: true
-            )
-            
-            try snapshotData?.write(to: snapshotURL)
-        } catch {
-            XCTFail("Failed to record snapshot with error: \(error)", file: file, line: line)
-        }
-    }
-    
-    private func makeSnapshotURL(named name: String, file: StaticString) -> URL {
-        // Ideally we should store the snapshots close to the test file so we can use the current file path which is `FeedSnapshotTests` file as the base path for our snapshots.
-        // Storing the snapshot to the same test directory also allows you to add it to git and share with others.
-        // Those snapshots should be stored and pushed to git so others can validate their changes against the stored snapshots.
-        return URL(fileURLWithPath: String(describing: file))
-            .deletingLastPathComponent()
-            .appendingPathComponent("snapshots")
-            .appendingPathComponent("\(name).png")
-    }
-    
-    // The `#file` holds the path to the current file: `FeedSnapshotTests.swift`.
-    private func makeSnapshotData(for snapshot: UIImage, file: StaticString = #file, line: UInt = #line) -> Data? {
-        guard let data = snapshot.pngData() else {
-            XCTFail("Failed to generate PNG data representation from snapshot", file: file, line: line)
-            return nil
-        }
-        
-        return data
-    }
-    
 }
 
 private extension FeedViewController {
